+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - [`utils.aria2`] `Downloader.get_feedback()` now includes overall `duration` and `speed`
 - Contents download now displays overall download stats with speed
+- Using offspot-config 2.0.0
+- Failed downloads clearly appear as such in logs
+- Checksum info for files (not base image) from config is passed to aria2 and checked
 
 ### Fixed